se esta reordenando todo el codigo. es posible que ahora no compile whisk3d. pero lo estoy volviendo una libreria que funciona perfectamente en el proyecto de tuludo (nombre que voy a cambiar porque ya alguien lo usa) estos cambios incluyen la opcion de poder importar modelos obj. aun sin textura. pero parece andar bien
