+++ Shared/variables.h
@@ -1,19 +1,3 @@
-//tipos de objetos
-enum { mesh, camera, light, empty, armature, curve };
-enum { Solid, MaterialPreview, Rendered };
-enum {
-	background,
-	blanco,
-	accent,
-	accentDark,
-	negro,
-	gris,
-	naranjaFace,
-	headerColor,
-	negroTransparente,
-	grisUI
-};
-
 enum { Constant, Linear, EaseInOut, EaseIn, EaseOut };
 
 enum { AnimPosition, AnimRotation, AnimScale };
@@ -30,11 +14,6 @@ enum{
 };
 
 enum { trackto, copyrotation, copylocation };
-
-//interpolacion
-enum {lineal, closest};
-
-enum{array, mirror,	screw, boolean};
 
 enum { pointLight, sunLight };
 
@@ -95,7 +74,6 @@ float FRUSTUM_FAR = 1000.f;     //far depth clipping plane
 int nextLightId = GL_LIGHT1;
 
 float angle = 55.0f;
-bool SimularZBuffer = false;
 bool ViewFromCameraActive = false;
 bool PodesCambiarElMalditoObjetoSeleccionado = false;
         
@@ -118,7 +96,6 @@ GLfloat PivotX = 0;
 GLfloat PivotY = 0;
 GLfloat PivotZ = 0;
 int CameraActive;
-int view = MaterialPreview;
 bool CameraToView = false;
 
 GLfloat LastRotX = 0;
@@ -142,7 +119,6 @@ bool PlayAnimation = false;
 int StartFrame = 1;
 int EndFrame = 250;
 int CurrentFrame = 1;
-bool redibujar = true; //solo redibuja si este valor esta en true
 bool showOrigins = true;
 bool show3DCursor = true;
 GLfloat Cursor3DposX = 0.0f;
@@ -158,18 +134,6 @@ int ShiftCount = 0;
 int valorRotacion = 0;
 
 //listados
-//RArray<Object> Objects;
-std::vector<Object> Objects;
-//RArray<Light> Lights;
-std::vector<Light> Lights;
-//RArray<Mesh> Meshes;
-std::vector<Mesh> Meshes;
-//RArray<TInt> Collection;
-std::vector<int> Collection;
-//RArray<Material> Materials;
-std::vector<Material> Materials;
-//RArray<Modifier> Modifiers;
-std::vector<Modifier> Modifiers;
 //RArray<SaveState> estadoObjetos;
 std::vector<SaveState> estadoObjetos;
 //RArray<Constraint> Constraints;
@@ -179,23 +143,7 @@ std::vector<AnimationObject> AnimationObjects;
 //RArray<ShapeKeyAnimation> ShapeKeyAnimations;
 std::vector<ShapeKeyAnimation> ShapeKeyAnimations;
 
-//RArray<TTexture> Textures;
-std::vector<Texture> Textures;
 int NumTexturasWhisk3D = 0;
-
-//array de colores
-GLfloat ListaColores[11][4] = {
-	{ 0.22, 0.28, 0.25, 1.0  },   //fondo
-	{ 1.0, 1.0, 1.0, 1.0     },   //blanco
-	{ 0.91, 0.50, 0.98, 1.0  },   //acento (violeta) 
-	{ 0.48, 0.30, 1.0, 1.0  },   //acento oscuro	
-	{ 0.0, 0.0, 0.0, 1.0     },   //negro
-	{ 0.12, 0.12, 0.12, 1.0  },   //gris
-	{ 0.94, 0.59, 0.17, 0.25f},   //naranja transparente
-	{ 0.22, 0.22, 0.22, 1.0  },    //cabezera de la barra de herramientas
-	{ 0.0, 0.0, 0.0, 0.25f     },   //negroTransparente
-	{ 0.757, 0.757, 0.757, 1.0     },   //azul de los iconos seleccionados
-};
 
 GLfloat LineaPiso[4]  =      { 0.22, 0.22, 0.22, 1.0 };
 GLfloat LineaPisoRoja[4]  =  { 0.56, 0.23, 0.28, 1.0 };
